Move step_register and step_uload_vmx from iso to common.
