+++ builder/vmware/common/step_register_test.go
@@ -1,8 +1,9 @@
-package iso
+package common
 
 import (
-	"github.com/mitchellh/multistep"
 	"testing"
+
+	"github.com/mitchellh/multistep"
 )
 
 func TestStepRegister_impl(t *testing.T) {
