+++ builder/vmware/common/step_upload_vmx.go
@@ -1,11 +1,11 @@
-package iso
+package common
 
 import (
 	"fmt"
+	"log"
 	"path/filepath"
 
 	"github.com/mitchellh/multistep"
-	vmwcommon "github.com/mitchellh/packer/builder/vmware/common"
 	"github.com/mitchellh/packer/packer"
 )
 
@@ -23,15 +23,16 @@ type StepUploadVMX struct {
 }
 
 func (c *StepUploadVMX) Run(state multistep.StateBag) multistep.StepAction {
-	driver := state.Get("driver").(vmwcommon.Driver)
+	driver := state.Get("driver").(Driver)
 
 	ui := state.Get("ui").(packer.Ui)
 	vmxPath := state.Get("vmx_path").(string)
 
 	if c.RemoteType == "esx5" {
-		remoteDriver, ok := driver.(vmwcommon.RemoteDriver)
+		remoteDriver, ok := driver.(RemoteDriver)
 		if ok {
 			remoteVmxPath := filepath.ToSlash(filepath.Join(fmt.Sprintf("%s", remoteDriver), filepath.Base(vmxPath)))
+			log.Printf("Uploading VMX file from %s to %s", vmxPath, remoteVmxPath)
 			if err := remoteDriver.Upload(remoteVmxPath, vmxPath); err != nil {
 				state.Put("error", fmt.Errorf("Error writing VMX: %s", err))
 				return multistep.ActionHalt
