+++ builder/vmware/vmx/builder.go
@@ -79,6 +79,13 @@ func (b *Builder) Run(ui packer.Ui, hook packer.Hook, cache packer.Cache) (packe
 			HTTPPortMin: b.config.HTTPPortMin,
 			HTTPPortMax: b.config.HTTPPortMax,
 		},
+		&vmwcommon.StepUploadVMX{
+			RemoteType: b.config.RemoteType,
+		},
+		&vmwcommon.StepRegister{
+			Format:         "foo",
+			KeepRegistered: false,
+		},
 		&vmwcommon.StepConfigureVNC{
 			VNCBindAddress:     b.config.VNCBindAddress,
 			VNCPortMin:         b.config.VNCPortMin,
